requirements.txt
flask
flask-cors
flask-login
flask-sqlalchemy
openai
python-dotenv
requests
twilio